requirements.txt
streamlit
google-generativeai
pandas
gspread
oauth2client
plotly
xlsxwriter